@@ -1,11 +1,13 @@
-const express = require("express");
-const app = express();
+const express = require('express')
+const app = express()
 
 // Heroku dynamically sets a port
-const PORT = process.env.PORT || 5000;
+const PORT = process.env.PORT || 5000
 
-app.use(express.static("dist"));
+app.use(express.static('dist'))
 
 app.listen(PORT, () => {
-  console.log("server started on port 5000");
-});
+
+
+  console.log('server started on port 5000')
+})
